+++ src/Firebase.php
@@ -6,6 +6,8 @@ namespace Kreait;
 
 use Kreait\Firebase\Auth;
 use Kreait\Firebase\Database;
+use Kreait\Firebase\Factory;
+use Kreait\Firebase\Firestore;
 use Kreait\Firebase\Messaging;
 use Kreait\Firebase\RemoteConfig;
 use Kreait\Firebase\Storage;
@@ -15,43 +17,17 @@ use Kreait\Firebase\Storage;
  */
 class Firebase
 {
-    /**
-     * @var Database
-     */
-    private $database;
-
-    /**
-     * @var Auth
-     */
-    private $auth;
-
-    /**
-     * @var Storage
-     */
-    private $storage;
-
-    /**
-     * @var RemoteConfig
-     */
-    private $remoteConfig;
-
-    /**
-     * @var Messaging
-     */
-    private $messaging;
+    /** @var Factory */
+    private $factory;
 
     /**
      * @internal
      *
      * @deprecated 4.33
      */
-    public function __construct(Database $database, Auth $auth, Storage $storage, RemoteConfig $remoteConfig, Messaging $messaging)
+    public function __construct(Factory $factory)
     {
-        $this->database = $database;
-        $this->auth = $auth;
-        $this->storage = $storage;
-        $this->remoteConfig = $remoteConfig;
-        $this->messaging = $messaging;
+        $this->factory = $factory;
     }
 
     /**
@@ -60,7 +36,7 @@ class Firebase
      */
     public function getDatabase(): Database
     {
-        return $this->database;
+        return $this->factory->createDatabase();
     }
 
     /**
@@ -69,7 +45,7 @@ class Firebase
      */
     public function getAuth(): Auth
     {
-        return $this->auth;
+        return $this->factory->createAuth();
     }
 
     /**
@@ -78,7 +54,7 @@ class Firebase
      */
     public function getStorage(): Storage
     {
-        return $this->storage;
+        return $this->factory->createStorage();
     }
 
     /**
@@ -87,7 +63,7 @@ class Firebase
      */
     public function getRemoteConfig(): RemoteConfig
     {
-        return $this->remoteConfig;
+        return $this->factory->createRemoteConfig();
     }
 
     /**
@@ -96,7 +72,15 @@ class Firebase
      */
     public function getMessaging(): Messaging
     {
+        return $this->factory->createMessaging();
+    }
 
-        return $this->messaging;
+    /**
+     * @deprecated 4.35 Use {@see \Kreait\Firebase\Factory::createFirestore()} instead
+     * @see \Kreait\Firebase\Factory::createFirestore()
+     */
+    public function getFirestore(): Firestore
+    {
+        return $this->factory->createFirestore();
     }
 }